Working status/command bar
Commands are now handled centrally by the status bar component. Other
components reigster their available commands with a mobx store and the
status bar component allows those commands, based on the keys from the
configuration.

It is now also possible to execute the commands by name by starting
with `:`. This means that in the future, not all commands must have a
shortcut key.

It should be easy to move the collection and its selected
ebvironment/request into mobx as well.