Errors and messages are now handled through mobx
There is a new messages store that tracks all messages. Instead of
throwing in the code, we now add new messages to the messages store and
gracefully handle `undefined` return values.

The messages in the messages store can be of different levels, like
loglevels.

The statusbar component has an mbox autorun useEffect hook which will
print any new messages to the status bar. Errors in red and warnings in
yellow.